Remove vtgate grpc tls flags
Not setting these flags is not an option if you want to connect using
TLS because `grpcclient.SecureDialOption` would add an `WithInsecure`
option.

Setting these flags is not an option if you want to connect using
a custom TLS configuraiton because the `grpcclient.SecureDialOption`
resulting option will be applied AFTER any user-provided
`WithCredentials` option.

See [1] for more context.

[1]: https://github.com/vitessio/vitess/issues/7053